+++ src/main/java/model/Account.java
@@ -1,12 +1,14 @@
 package model;
 
+import java.math.BigInteger;
+
 public interface Account {
 
     public void deposit(double amount);
 
     public void withdraw(double amount);
 
-    public double getBalance();
+    public BigInteger getBalance();
 
     public int getAccountId();
